refactor(roller): move generic roller into heroRoller subsystem
