+++ module/herosystem6e.mjs
@@ -16,20 +16,20 @@ import { CreateHeroCompendiums } from "./heroCompendiums.mjs";
 import { setPerceptionModes } from "./herovision/vision.mjs";
 import { HeroPointVisionSource } from "./herovision/vision.mjs";
 import { EffectsPanel } from "./effects-panel.mjs";
-import { GenericRoller } from "./genericRoller.mjs";
 import { HeroSocketHandler } from "./heroSocketHandler.mjs";
 import { HeroSystem6eChatMessage } from "./heroChatMessage.mjs";
 
 import { HeroSystem6eActor } from "./actor/actor.mjs";
-import { HeroSystemActorSheet } from "./actor/actor-sheet.mjs";
-import { HeroSystemActorSheetV2 } from "./applications/actor/actor-sheet-v2.mjs";
-import { HeroSystemActorSavuoriSheet } from "./actor/actor-savuori-sheet.mjs";
-import { HeroSystem6eToken, HeroSystem6eTokenDocument } from "./token/actor-token.mjs";
 import {
     HeroSystem6eActorActiveEffects,
     HeroSystem6eActorActiveEffectsSystemData,
 } from "./actor/actor-active-effects.mjs";
 import { HeroSystemActiveEffectConfig } from "./actor/active-effect-config.mjs";
+import { HeroSystemActorSheet } from "./actor/actor-sheet.mjs";
+import { HeroSystemActorSheetV2 } from "./applications/actor/actor-sheet-v2.mjs";
+import { HeroSystemActorSavuoriSheet } from "./actor/actor-savuori-sheet.mjs";
+
+import { GenericRoller } from "./heroRoller/genericRoller.mjs";
 
 import { HeroSystemItemSheetV2 } from "./applications/item/item-sheet-v2.mjs";
 import { ItemVppConfig } from "./applications/apps/item-vpp-config.mjs";
@@ -55,6 +55,7 @@ import { testingMockRollInitialize } from "./testing/dice-testing-helper.mjs";
 import "./testing/testing-main.mjs";
 import { HeroSystem6eEndToEndTest } from "./testing/end-to-end.mjs";
 
+import { HeroSystem6eToken, HeroSystem6eTokenDocument } from "./token/actor-token.mjs";
 import { HeroSystemTokenHud } from "./token/heroSystemTokenHud.mjs";
 
 import { expireEffects } from "./utility/util.mjs";
